hw5/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.14.0
psutil>=5.9.0
pandas>=2.0.0
onnxruntime-gpu>=1.16.0
tabulate>=0.9.0
openvino>=2023.1.0
onnx>=1.14.0 